fix: 시큐리티 작동 제외, 패키지 위치 수정
