+++ src/main/java/Iniro/kTrip/domain/Favorite.java
@@ -1,4 +1,4 @@
-package domain;
+package Iniro.kTrip.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,24 +6,20 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "review")
+@Entity(name="Favorite")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Review {
-
+public class Favorite {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int review_id;
+    private int fav_id;
 
     @ManyToOne
     @JoinColumn(name="member_id")
     private Member member;
 
-    @Column
-    private int point;
-
-    @Column(length = 255)
-    private String content;
+    @Column(length = 20)
+    private String region;
 }
