+++ src/main/java/Iniro/kTrip/domain/Review.java
@@ -1,4 +1,4 @@
-package domain;
+package Iniro.kTrip.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,20 +6,24 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity(name="Favorite")
+@Entity(name = "review")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Favorite {
+public class Review {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int fav_id;
+    private int review_id;
 
     @ManyToOne
     @JoinColumn(name="member_id")
     private Member member;
 
-    @Column(length = 20)
-    private String region;
+    @Column
+    private int point;
+
+    @Column(length = 255)
+    private String content;
 }
