+++ src/main/java/Iniro/kTrip/token/JwtService.java
@@ -1,7 +1,7 @@
-package token;
+package Iniro.kTrip.token;
 
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
-import org.springframework.security.oauth2.jwt.Jwt;
+//import org.springframework.security.oauth2.jwt.Jwt;
 
 import java.time.Duration;
 import java.util.Date;
