+++ src/test/java/repository/MemberRepositoryTest.java
@@ -1,14 +1,5 @@
 package repository;
 
-import domain.Member;
-import dto.MemberDto;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-
 
 //class MemberRepositoryTest {
 //
